refactor(react-langchain): split runtime file and adopt createRuntimeExtras
decompose the 505-line useStreamRuntime.ts into the standard adapter layout:
accessor/action hooks move to hooks.ts, option/extras types to types.ts, the
outbound AppendMessage mapper to convertMessages.ts, and the runtime file is
left to orchestrate only. the per-package Symbol + asLangChainRuntimeExtras
guard is replaced by the shared createRuntimeExtras helper from
@assistant-ui/core/internal. behavior is unchanged; tests updated for the new
module paths and the helper's error message.
