+++ .changeset/react-langchain-runtime-extras.md
@@ -1,0 +1,5 @@
+---
+"@assistant-ui/react-langchain": patch
+---
+
+refactor: split the useStreamRuntime file into hooks/types/converter modules and adopt the shared createRuntimeExtras helper
