+++ packages/react-langchain/src/__tests__/langChainTestUtils.ts
@@ -2,14 +2,14 @@ import type { Mock } from "vitest";
 
 export type Selector = (state: unknown) => unknown;
 
-// The real guard symbol in useStreamRuntime is module-private. To stand in for
+// The createRuntimeExtras brand symbol is module-private. To stand in for
 // runtime-produced extras, wrap a plain object in a Proxy whose `has` trap
-// claims any symbol described "langchain-runtime-extras" is present.
+// claims any symbol described "useStreamRuntime extras" is present.
 export const makeExtras = (extras: Record<string, unknown>) =>
   new Proxy(extras, {
     has: (target, key) =>
       (typeof key === "symbol" &&
-        key.description === "langchain-runtime-extras") ||
+        key.description === "useStreamRuntime extras") ||
       Reflect.has(target, key),
   });
