+++ packages/react-langchain/src/runtimeExtras.ts
@@ -1,0 +1,5 @@
+import { createRuntimeExtras } from "@assistant-ui/core/internal";
+import type { LangChainRuntimeExtras } from "./types";
+
+export const langChainExtras =
+  createRuntimeExtras<LangChainRuntimeExtras>("useStreamRuntime");
