+++ packages/react-langchain/src/types.ts
@@ -1,4 +1,15 @@
-import type { MessageStatus } from "@assistant-ui/core";
+import type {
+  AttachmentAdapter,
+  DictationAdapter,
+  ExternalStoreSharedOptions,
+  FeedbackAdapter,
+  MessageStatus,
+  RealtimeVoiceAdapter,
+  RemoteThreadListAdapter,
+  SpeechSynthesisAdapter,
+} from "@assistant-ui/core";
+import type { AssistantCloud } from "assistant-cloud";
+import type { UseStreamOptions, AssembledToolCall } from "@langchain/react";
 
 /** Known content block types from @langchain/core messages. */
 export type LangChainContentBlock =
@@ -47,3 +58,68 @@ export type LangChainBaseMessage = {
   /** Present on ToolMessage */
   artifact?: unknown;
 };
+
+export type LangChainRuntimeExtras = {
+  interrupt: { value?: unknown } | undefined;
+  interrupts: readonly { value?: unknown }[];
+  toolCalls: readonly AssembledToolCall[];
+  error: unknown;
+  submit: (
+    values: Record<string, unknown> | null | undefined,
+    options?: Record<string, unknown>,
+  ) => Promise<void>;
+  respond: (
+    response: unknown,
+    options?: Record<string, unknown>,
+  ) => Promise<void>;
+  respondAll: (
+    responsesById: Record<string, unknown>,
+    options?: Record<string, unknown>,
+  ) => Promise<void>;
+  values: Record<string, unknown>;
+  messagesKey: string;
+};
+
+export type LangChainRuntimeExtraOptions = ExternalStoreSharedOptions & {
+  cloud?: AssistantCloud | undefined;
+  adapters?:
+    | {
+        attachments?: AttachmentAdapter | undefined;
+        speech?: SpeechSynthesisAdapter | undefined;
+        dictation?: DictationAdapter | undefined;
+        voice?: RealtimeVoiceAdapter | undefined;
+        feedback?: FeedbackAdapter | undefined;
+      }
+    | undefined;
+  /**
+   * When the user sends a new message while previous tool calls are
+   * still pending, automatically submit `tool` messages that cancel
+   * them so the agent's tool-call accounting stays consistent.
+   * Defaults to `true`.
+   */
+  autoCancelPendingToolCalls?: boolean | undefined;
+  /**
+   * Routes the Cancel button's click to `useStream().stop()`. On by
+   * default. Pass `false` to disable the Cancel button.
+   */
+  unstable_allowCancellation?: boolean | undefined;
+  /**
+   * Custom `RemoteThreadListAdapter`. When provided, replaces the
+   * cloud-backed thread list adapter.
+   */
+  unstable_threadListAdapter?: RemoteThreadListAdapter | undefined;
+  /** Custom thread-creation hook, forwarded to the cloud adapter. */
+  create?: (() => Promise<{ externalId: string | undefined }>) | undefined;
+  /** Custom thread-deletion hook, forwarded to the cloud adapter. */
+  delete?: ((threadId: string) => Promise<void>) | undefined;
+};
+
+// Distribute the intersection through the union arms of `UseStreamOptions`
+// (`AgentServerOptions | CustomAdapterOptions`). Writing `UseStreamOptions & X`
+// directly collapses arm tracking, so `Omit<…, "cloud">` and the like would
+// produce a flattened structural type that no longer matches either arm.
+export type UseStreamRuntimeOptions = UseStreamOptions extends infer O
+  ? O extends UseStreamOptions
+    ? O & LangChainRuntimeExtraOptions
+    : never
+  : never;
